Update install scripts (#181)
* updates the redeploy script to use standard systemctl command for managing tomcat
* updates installation script to assign permissions on all the created directories to tomcat user
* updates the permissions of darshini-es directory without the new permissions, 
  elastic search fails to save data and logs; elastic search throws up an error.

+++ scripts/install.sh
@@ -14,7 +14,9 @@ yes | sudo apt-get -y --force-yes install git
 
 
 #install maven
-curl -O http://www-us.apache.org/dist/maven/maven-3/3.5.2/binaries/apache-maven-3.5.2-bin.tar.gz
+if [ ! -f apache-maven-3.5.2-bin.tar.gz ]; then
+    curl -O http://www-us.apache.org/dist/maven/maven-3/3.5.2/binaries/apache-maven-3.5.2-bin.tar.gz
+fi
 tar -xzf apache-maven-3.5.2-bin.tar.gz
 sudo mv apache-maven-3.5.2 /opt/maven 
 
@@ -48,7 +50,9 @@ sudo groupadd tomcat
 sudo useradd -s /bin/false -g tomcat -d /opt/tomcat tomcat
 
 #download and install tomcat
-curl -O http://www-eu.apache.org/dist/tomcat/tomcat-8/v8.5.27/bin/apache-tomcat-8.5.27.tar.gz
+if [ ! -f apache-tomcat-8.5.29.tar.gz ]; then
+    curl -O http://redrockdigimark.com/apachemirror/tomcat/tomcat-8/v8.5.29/bin/apache-tomcat-8.5.29.tar.gz
+fi
 sudo mkdir /opt/tomcat
 sudo tar xzf apache-tomcat-8*tar.gz -C /opt/tomcat --strip-components=1
 
@@ -58,22 +62,29 @@ sudo chown -R tomcat:tomcat /opt/tomcat
 #give the tomcat group read access to the conf directory and all of its contents, and execute access to the directory itself
 sudo cp conf/tomcat.service /etc/systemd/system/
 sudo cp conf/tomcat-users.xml /opt/tomcat/conf/tomcat-users.xml
+sudo chown tomcat:tomcat /opt/tomcat/conf/tomcat-users.xml
+
+#make tomcat a part of start-up scripts
+sudo update-rc.d tomcat defaults
+sudo update-rc.d tomcat enable
 
 #reload the systemd daemon so that it knows about our service file
 sudo systemctl daemon-reload
 sudo systemctl start tomcat
 
-#copy the correct property files and create the required directories
+#copy the correct property files
 cp -f conf/*.properties src/main/resources/META-INF/
+
+#create directories for storing elastic search data and logs
 sudo mkdir -p /opt/darshini-es/data
-sudo chmod 777 /opt/darshini-es/data
 sudo mkdir -p /opt/darshini-es/logs
-sudo chmod 777 /opt/darshini-es/logs
+sudo chown -R tomcat:tomcat /opt/darshini-es
+sudo chmod -R 777 /opt/darshini-es
+
+#create log file for darshini
 sudo mkdir -p /opt/darshini-logs
-sudo chmod 777 /opt/darshini-logs
 sudo touch /opt/darshini-logs/darshini
-sudo chown tomcat:tomcat /opt/darshini-logs/darshini
-sudo chmod 777 /opt/darshini-logs/darshini
+sudo chown -R tomcat:tomcat /opt/darshini-logs
 
 #install nodejs and npm
 curl -sL https://deb.nodesource.com/setup_8.x -o nodesource_setup.sh
